add new labeler for datasets
