--- extractive_labeler/data/summ1.txt
@@ -1,1 +1,0 @@
-manchester united flop anderson was sent off for internacional this week\nanderson saw red for an off - the - ball shove during the first half of 1 - 1 draw\nteam - mate fabricio stole the limelight by swearing at his own fans
