+++ extractive_labeler/sample_in.json
@@ -1,0 +1,1 @@
+{"doc": "things have n't been going well for manchester united flop anderson since he left old trafford to return to brazil , but at least on wednesday his own errors were outshone by the stupidity of a team - mate\nthe brazil midfielder was sent off for internacional after a first - half off - the - ball shove on ypiranga erechim striker neto that also saw his opponent sent off for retaliating with an elbow\nbut fortunately for anderson , his colleague fabricio stole the limelight by swearing at his own fans and becoming the third player to be shown the red card , reducing internacional to nine men in the 1 - 1 draw in porto alegre\nformer manchester united midfielder anderson shoved neto off the ball to earn a red card neto also saw red after he retaliated by by appearing to aim an elbow at anderson in porto alegre the two players listen to the referee as he reaches for his card from his back pocket on wednesday night anderson shows his shock to be shown the red card for what he believed to be an innocent challenge while some players offered their hands in apology to the fans after being repeatedly booed , fabricio took a different route and showed two middle fingers to the crowd leaving referee luis teixeira rocha with no choice but to give him his marching orders\nthe 28 - year - old left back reacted by tearing off his shirt , throwing it to the floor and shouting ' i 'm leaving , i 'm leaving , ' as he walked down the tunnel\nhe has already been suspended until the end of the week by his club\nanderson ended his seven - and - a - half year association with manchester united in february by signing a four - year deal with internacional , but his return to brazil has not gone as planned\ninternacional 's fabricio was also sent off later in the game for aiming his middle fingers at booing fans fabricio threw his shirt to the ground after being shown the red card and shouted ' i 'm leaving , i 'm leaving ' having missed a penalty on his debut , he was substituted later in february after just 36 minutes needing oxygen while playing in the high - altitude estadio hernando siles in la paz\nwhile internacional sit top of the table in the campeonato gaucho , anderson has faced a familiar struggle for fitness despite making eight appearances since his move\nanderson ended a miserable seven - and - a - half year spell with manchester united by leaving this season", "summaries": "manchester united flop anderson was sent off for internacional this week\nanderson saw red for an off - the - ball shove during the first half of 1 - 1 draw\nteam - mate fabricio stole the limelight by swearing at his own fans"}
